builder/auth-scraper/reproduce_issue.py

try:
    from nacl.signing import SigningKey
    import base64

    # Simulate a base58 encoded key (just random bytes encoded to base58-like string for testing if library handles string)
    # Actually, pynacl SigningKey expects bytes. 
    # If I pass a string, it should fail.
    
    try:
        k = SigningKey("notbytes")
        print("Success with string")
    except Exception as e:
        print(f"Failed with string: {e}")

    # If the user intended to support base58, they probably need 'base58' library or similar.
    try:
        import base58
        print("base58 library is available")
    except ImportError:
        print("base58 library is NOT available")

except ImportError:
    print("pynacl not installed")
